+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -2,11 +2,11 @@ package com.techcourse;
 
 import jakarta.servlet.ServletContext;
 import nextstep.mvc.DispatcherServlet;
-import nextstep.mvc.handleradaptor.HandlerAdapterRegistry;
-import nextstep.mvc.handleradaptor.HandlerExecutionHandlerAdapter;
-import nextstep.mvc.handleradaptor.ManualHandlerAdaptor;
-import nextstep.mvc.handlermapping.AnnotationHandlerMapping;
-import nextstep.mvc.handlermapping.HandlerMappingRegistry;
+import nextstep.mvc.HandlerAdapterRegistry;
+import nextstep.mvc.controller.tobe.HandlerExecutionHandlerAdapter;
+import nextstep.mvc.controller.asis.ControllerHandlerAdaptor;
+import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
+import nextstep.mvc.HandlerMappingRegistry;
 import nextstep.web.WebApplicationInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -24,7 +24,7 @@ public class AppWebApplicationInitializer implements WebApplicationInitializer {
         dispatcherServlet.addHandlerMapping(new ManualHandlerMapping());
         dispatcherServlet.addHandlerMapping(new AnnotationHandlerMapping("com.techcourse.controller"));
         dispatcherServlet.addHandlerAdapter(new HandlerExecutionHandlerAdapter());
-        dispatcherServlet.addHandlerAdapter(new ManualHandlerAdaptor());
+        dispatcherServlet.addHandlerAdapter(new ControllerHandlerAdaptor());
 
         final var dispatcher = servletContext.addServlet("dispatcher", dispatcherServlet);
         dispatcher.setLoadOnStartup(1);
