+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -4,10 +4,6 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.handleradaptor.HandlerAdapter;
-import nextstep.mvc.handleradaptor.HandlerAdapterRegistry;
-import nextstep.mvc.handlermapping.HandlerMapping;
-import nextstep.mvc.handlermapping.HandlerMappingRegistry;
 import nextstep.mvc.view.ModelAndView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
