+++ mvc/src/main/java/nextstep/mvc/HandlerAdapter.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.handleradaptor;
+package nextstep.mvc;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
