+++ mvc/src/main/java/nextstep/mvc/controller/tobe/HandlerExecutionHandlerAdapter.java
@@ -1,8 +1,8 @@
-package nextstep.mvc.handleradaptor;
+package nextstep.mvc.controller.tobe;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.tobe.HandlerExecution;
+import nextstep.mvc.HandlerAdapter;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
