+++ mvc/src/test/java/nextstep/mvc/HandlerAdapterRegistryTest.java
@@ -1,11 +1,13 @@
-package nextstep.mvc.handleradaptor;
+package nextstep.mvc;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.lang.reflect.InvocationTargetException;
+import nextstep.mvc.controller.asis.ControllerHandlerAdaptor;
 import nextstep.mvc.controller.tobe.HandlerExecution;
+import nextstep.mvc.controller.tobe.HandlerExecutionHandlerAdapter;
 import org.junit.jupiter.api.Test;
 import samples.TestController;
 
@@ -17,9 +19,9 @@ class HandlerAdapterRegistryTest {
         // given
         final HandlerAdapterRegistry handlerAdapterRegistry = new HandlerAdapterRegistry();
         final HandlerExecutionHandlerAdapter handlerExecutionHandlerAdapter = new HandlerExecutionHandlerAdapter();
-        final ManualHandlerAdaptor manualHandlerAdaptor = new ManualHandlerAdaptor();
+        final ControllerHandlerAdaptor controllerHandlerAdaptor = new ControllerHandlerAdaptor();
         handlerAdapterRegistry.addHandlerAdapter(handlerExecutionHandlerAdapter);
-        handlerAdapterRegistry.addHandlerAdapter(manualHandlerAdaptor);
+        handlerAdapterRegistry.addHandlerAdapter(controllerHandlerAdaptor);
 
         final HandlerExecution handlerExecution = new HandlerExecution(
                 TestController.class.getConstructor().newInstance(),
