+++ mvc/src/test/java/nextstep/mvc/controller/asis/ControllerHandlerAdaptorTest.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.handleradaptor;
+package nextstep.mvc.controller.asis;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
@@ -7,24 +7,24 @@ import static org.mockito.Mockito.when;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.Map;
-import nextstep.mvc.controller.asis.ForwardController;
+import nextstep.mvc.HandlerAdapter;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 import org.assertj.core.api.AssertionsForClassTypes;
 import org.junit.jupiter.api.Test;
 import samples.TestController;
 
-class ManualHandlerAdaptorTest {
+class ControllerHandlerAdaptorTest {
 
     @Test
     void ManualHandlerAdaptor는_Controller_인터페이스를_상속받은_클래스의_객체를_처리할_수_있다() {
         // given
         final ForwardController forwardController = new ForwardController("/123.html");
 
-        final ManualHandlerAdaptor manualHandlerAdaptor = new ManualHandlerAdaptor();
+        final ControllerHandlerAdaptor controllerHandlerAdaptor = new ControllerHandlerAdaptor();
 
         // when
-        final boolean result = manualHandlerAdaptor.supports(forwardController);
+        final boolean result = controllerHandlerAdaptor.supports(forwardController);
 
         // then
         assertThat(result).isTrue();
@@ -35,10 +35,10 @@ class ManualHandlerAdaptorTest {
         // given
         final TestController testController = new TestController();
 
-        final ManualHandlerAdaptor manualHandlerAdaptor = new ManualHandlerAdaptor();
+        final ControllerHandlerAdaptor controllerHandlerAdaptor = new ControllerHandlerAdaptor();
 
         // when
-        final boolean result = manualHandlerAdaptor.supports(testController);
+        final boolean result = controllerHandlerAdaptor.supports(testController);
 
         // then
         assertThat(result).isFalse();
@@ -47,7 +47,7 @@ class ManualHandlerAdaptorTest {
     @Test
     void Handler를_처리하면_modelAndView를_반환한다() throws Exception {
         // given
-        final HandlerAdapter handlerAdapter = new ManualHandlerAdaptor();
+        final HandlerAdapter handlerAdapter = new ControllerHandlerAdaptor();
 
         final var request = mock(HttpServletRequest.class);
         final var response = mock(HttpServletResponse.class);
